Version 1.2.0 (may 17th)
this version contains a new set of files that allow for the addition of
triangles (right-angled) (rotated 90 degrees for 4 rotated copies) the
triangles and rectangles were all done by aakash and not the person who
created the physics. There are some bugs especially if the ball enters a
corner, but that can be overseen as there is no easy or worthwhile
solution.

To add an 'solid' object in the program you must check for it in the
MoveEngine class (there is a specific method that allows you to do this,
one for triangles and one for rectangles). You must then create a visual
representation of the object in the Main class using graphics2d shapes.
NOTE: triangles will need drawPolygon, or a class of the sort.